requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
google-generativeai
pydantic
# Später für die Datenbank:
# sqlalchemy
# psycopg2-binary